+++ Opcodes/ugens9.h
@@ -27,7 +27,7 @@ typedef struct {
     OPDS    h;
     MYFLT   *ar1,*ar2,*ar3,*ar4,*ain,*ifilno,*channel;
     MEMFIL  *mfp;
-    long    Hlen, Hlenpadded,incount,outcnt,obufsiz;
+    int32    Hlen, Hlenpadded,incount,outcnt,obufsiz;
     int     nchanls; /* number of channels we are actually processing */
     MYFLT   *H,*cvlut,*outhead,*outail,*obufend;
     AUXCH   auxch;    /* use AUXDS to manage the following buffer spaces */
@@ -43,24 +43,24 @@ typedef struct {
 typedef struct {
     OPDS    h;
     MYFLT   *ar1, *ar2, *ar3, *ar4, *ain,*ifilno,*partitionSize,*channel;
-    long    numPartitions;
-    long    Hlen, Hlenpadded;
+    int32    numPartitions;
+    int32    Hlen, Hlenpadded;
     int     nchanls;    /* number of channels we are actually processing */
 
     AUXCH   H;                  /* array of Impulse Responses */
 
     AUXCH   savedInput; /* the last Hlen input samps for overlap-save method */
-    long    inCount;    /* index to write to savedInput */
+    int32   inCount;    /* index to write to savedInput */
 
     AUXCH   workBuf;    /* work buf for current partion convolution */
     MYFLT   *workWrite; /* current index for writing input samps */
 
     AUXCH   convBuf;    /* circular buf accumulating partitioned convolutions */
-    long    curPart;    /* "current" segment in convBuf */
+    int32   curPart;    /* "current" segment in convBuf */
 
     AUXCH   output;             /* circular buf accumulating output samples */
-    long    outBufSiz;  /* hlenpadded or 2*ksmps, whichever is greater */
+    int32   outBufSiz;  /* hlenpadded or 2*ksmps, whichever is greater */
     MYFLT   *outWrite, *outRead; /* i/o pointers to the output buf */
-    long    outCount;   /* number of valid samples in the outbuf */
+    int32   outCount;   /* number of valid samples in the outbuf */
 } PCONVOLVE;
